Move Enterable adapter to _adapters module.
